indexer: remove bindings
Removes the dependency on the `op-bindings/bindings` package.
This is to improve devex in the monorepo and reduce CI time
as maintaining the bindings autogenerated in each PR just
doesn't scale. Now each service is responsible for their own
bindings. In the future, we can work towards releases of
the bindings when contracts are released.

Follows:
- https://github.com/ethereum-optimism/optimism/pull/10213
- https://github.com/ethereum-optimism/optimism/pull/10218
- https://github.com/ethereum-optimism/optimism/pull/10221